Use LintianIssue API.

+++ fixers/debian-pyversions-is-obsolete.py
@@ -2,7 +2,7 @@
 import os
 import sys
 
-from lintian_brush.fixer import report_result, fixed_lintian_tag
+from lintian_brush.fixer import report_result, LintianIssue
 
 if not os.path.exists('debian/pyversions'):
     sys.exit(2)
@@ -11,7 +11,9 @@ with open('debian/pyversions', 'r') as f:
     pyversions = f.read().strip()
 
 if pyversions.startswith('2.'):
-    os.unlink('debian/pyversions')
-    fixed_lintian_tag('source', "debian-pyversions-is-obsolete")
+    issue = LintianIssue('source', "debian-pyversions-is-obsolete")
+    if issue.should_fix():
+        os.unlink('debian/pyversions')
+        issue.report_fixed()
 
 report_result("Remove obsolete debian/pyversions.")
